changing main to typescript
